@@ -2,8 +2,9 @@ package trace.event.putao.com.scraprecyclerview;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 public class ScrapTestActivity extends AppCompatActivity {
 
