@@ -1,9 +1,10 @@
 package trace.event.putao.com.scraprecyclerview;
 
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 /**
  * Created by riven_chris on 2018/3/26.
